+++ x/sao/types/message_complete.go
@@ -9,7 +9,7 @@ const TypeMsgComplete = "complete"
 
 var _ sdk.Msg = &MsgComplete{}
 
-func NewMsgComplete(creator string, orderId uint64, cid string, size int32) *MsgComplete {
+func NewMsgComplete(creator string, orderId uint64, cid string, size uint64) *MsgComplete {
 	return &MsgComplete{
 		Creator: creator,
 		OrderId: orderId,
